move model-nav-list
