+++ src/app/pages/right-tab-nav/model-nav-list/model-nav-list.component.scss
