+++ src/app/pages/right-tab-nav/right-tab-nav.component.ts
@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { ModelNavListComponent } from "../../shared/components/model-nav-list/model-nav-list.component";
+import { ModelNavListComponent } from "./model-nav-list/model-nav-list.component";
 import { MatTabsModule } from '@angular/material/tabs';
 import { MeasureTabsComponent } from "./measure-tabs/measure-tabs.component";
 import { CrossSectionTabComponent } from "./cross-section-tab/cross-section-tab.component";
